requirements.txt
pandas
scikit-learn
torch
simpy
matplotlib
requests
python-dotenv
